% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_classes.R
\name{available_classes}
\alias{available_classes}
\title{Available Classes}
\usage{
available_classes(filtration = NULL, input = NULL)
}
\arguments{
\item{filtration}{Character string specifying filtration}

\item{input}{Character string specifying input type}
}
\value{
Character vector that has the available classes
}
\description{
Lists the classes supported by a filtration and input type
}
