% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_filtrations.R
\name{available_filtrations}
\alias{available_filtrations}
\title{Available Filtrations}
\usage{
available_filtrations()
}
\value{
Dataframe with one row per filtration and a column with supported
engines
}
\description{
Lists all filtrations supported by each engine
}
