% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_inputs.R
\name{available_inputs}
\alias{available_inputs}
\title{Available Inputs}
\usage{
available_inputs(filtration = NULL)
}
\arguments{
\item{filtration}{Character string specifying filtration}
}
\value{
Character vector that has the available input types
}
\description{
Lists the input types supported by a filtration
}
