% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_check.R
\name{check_packages}
\alias{check_packages}
\title{function to check if packages are installed}
\usage{
check_packages(object)
}
\description{
This function prints an error when the user does not have the
required packages installed. It directs the user to install the packages.
}
