% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_double_dispatch_solution.R
\docType{data}
\name{filtration_type}
\alias{filtration_type}
\title{include aaa.r}
\format{
An object of class \code{S7_property} of length 6.
}
\usage{
filtration_type
}
\description{
include aaa.r
}
\keyword{datasets}
