% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{new_property}
\alias{new_property}
\title{importFrom S7 new_property}
\usage{
new_property(
  class = class_any,
  getter = NULL,
  setter = NULL,
  validator = NULL,
  default = NULL,
  name = NULL
)
}
\description{
importFrom S7 new_property
}
