% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_classes.R
\name{PH}
\alias{PH}
\alias{PH_pointcloud}
\alias{PH_raster}
\title{Classes for persistent homology specifications}
\usage{
PH(
  filtration = "vietoris_rips",
  engine = "TDA",
  library = NA_character_,
  max_dimension = 1
)

PH_pointcloud(
  filtration = "vietoris_rips",
  engine = "TDA",
  library = NA_character_,
  max_dimension = 1,
  max_diameter = NA_real_
)

PH_raster(
  filtration = "vietoris_rips",
  engine = "TDA",
  library = NA_character_,
  max_dimension = 1,
  max_scale = NA_real_,
  sublevel = TRUE
)
}
\arguments{
\item{filtration}{character; filtration used in persistence calculation}

\item{engine}{character; back-end engine used in persistence calculation}

\item{library}{character; c++ library used for TDA engine}

\item{max_dimension}{character; maximum homological dimension to compute persistence}

\item{max_diameter}{character; maximum threshold for rips filtration (point clouds)}

\item{max_scale}{character; maximum threshold for rips filtration (rasters)}

\item{sublevel}{boolean; specifies sublevel or superlevel filtration}
}
\value{
S7 object storing user specification for calculating persistence
}
\description{
These class define the parameters needed for persistence
calculations and allow for reuse of specification for different data sets.
The parent class, \code{PH}, stores core attributes that are common among all
persistence calculations. The point cloud subclass is used to handle
distance matrices and point cloud arrays. The raster subclass is used
to handle grid data encoded as matrices or arrays.
}
