% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend_engine.R
\name{recommend_engine}
\alias{recommend_engine}
\title{Recommend an Engine}
\usage{
recommend_engine(filtration)
}
\arguments{
\item{filtration}{Character string specifying filter}
}
\value{
Character string that has the recommended engine
}
\description{
Recommends an engine based on filtration
}
