% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_engines.R
\name{available_engines}
\alias{available_engines}
\title{Available Engines}
\usage{
available_engines(filtration)
}
\arguments{
\item{filtration}{Character string specifying filter}
}
\value{
Character vector that has the available engines
}
\description{
Lists the engines available for a filtration
}
