% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_double_dispatch_solution.R
\name{compute_persistence}
\alias{compute_persistence}
\title{Compute Persistent Homology}
\usage{
compute_persistence(object, data, ...)
}
\arguments{
\item{object}{user specification}

\item{data}{object on which to compute persistent homology}
}
\value{
class \code{persistence} object
}
\description{
This function is an S7 generic which dispatches based on
the user specification and the class of the data. The function standardizes
output by converting it to a \code{persistence} object.
}
\examples{

obj <- PH_pointcloud(filtration = "vietoris_rips",
              engine = "ripserr",
              max_dimension = 1,
              max_diameter = 1)
data <- matrix(rnorm(30), ncol = 3)
compute_persistence(obj, data)

obj <- PH_raster(filtration = "cubical",
          engine = "TDA",
          library = "GUDHI",
          max_dimension = 2,
          max_scale = 10,
          sublevel = TRUE)
data <- volcano
compute_persistence(obj, data)
}
