% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_persistence.R,
%   R/current_double_dispatch_solution.R
\name{compute_persistence}
\alias{compute_persistence}
\title{Compute Persistent Homology}
\usage{
compute_persistence(object, data, ...)

compute_persistence(object, data, ...)
}
\arguments{
\item{object}{user specification}

\item{data}{object on which to compute persistent homology}

\item{...}{additional engine-specific arguments}
}
\value{
class \code{persistence} object

class \code{persistence} object
}
\description{
This function is an S7 generic which dispatches based on
the user specification and the class of the data. The function standardizes
output by converting it to a \code{persistence} object.

This function is an S7 generic which dispatches based on
the user specification and the class of the data. The function standardizes
output by converting it to a \code{persistence} object.
}
