% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Persistence_classes.R
\name{Persistence}
\alias{Persistence}
\alias{PersistencePointCloud}
\alias{PersistenceGrid}
\title{Classes for persistent homology specifications}
\usage{
Persistence(engine = "TDA", library = NA_character_, max_dimension = 1)

PersistencePointCloud(
  engine = "TDA",
  library = NA_character_,
  max_dimension = 1,
  filtration = "vietoris_rips",
  max_diameter = NA_real_
)

PersistenceGrid(
  engine = "TDA",
  library = NA_character_,
  max_dimension = 1,
  filtration = "cubical",
  max_scale = NA_real_,
  sublevel = TRUE
)
}
\arguments{
\item{engine}{character; back-end engine used in persistence calculation}

\item{library}{character; c++ library used for TDA engine}

\item{max_dimension}{character; maximum homological dimension to compute
persistence}

\item{filtration}{character; filtration used in persistence calculation}

\item{max_diameter}{character; maximum threshold for rips filtration
(point clouds)}

\item{max_scale}{character; maximum threshold for rips filtration (grids)}

\item{sublevel}{boolean; specifies sublevel or superlevel filtration}
}
\value{
S7 object storing user specification for calculating persistence
}
\description{
These class define the parameters needed for persistence
calculations and allow for reuse of specification for different data sets.
The parent class, \code{Persistence}, stores core attributes that are common
among all persistence calculations. The point cloud subclass is used to
handle distance matrices and point cloud arrays. The grid subclass is
used to handle matrix or array data treated as function values over a
domain.
}
