% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend_engine.R
\name{recommend_engine}
\alias{recommend_engine}
\title{Recommend Engine}
\usage{
recommend_engine(filtration, input)
}
\arguments{
\item{filtration}{Character string specifying filtration}

\item{input}{Character string specifying input type}
}
\value{
Character string with the recommend engine
}
\description{
Returns the recommended engine for a filtration and input type
}
